requirements.txt
kazoo>=1.2
requests
flask
PyYaml
